image-compression-service/image-compressor-lambda/requirements.txt
altgraph==0.17.2
boto3==1.28.3
botocore==1.31.3
future==0.18.2
jmespath==1.0.1
macholib==1.15.2
Pillow==10.0.0
pip==23.2
python-dateutil==2.8.2
s3transfer==0.6.1
setuptools==58.0.4
six==1.15.0
urllib3==1.26.16
wheel==0.37.0
